requirements.txt
numpy
pyaudio
pydub
simpleaudio
synthesizer
tkinter
ttktheme
